progress

+++ libs/vision/include/mrpt/vision/CDifodo.h
@@ -189,7 +189,7 @@ class CDifodo
 
    public:
 	/** Frames per second (Hz) */
-	float fps;
+	double fps;
 
 	/** Resolution of the images taken by the range camera */
 	unsigned int cam_mode;  // (1 - 640 x 480, 2 - 320 x 240, 4 - 160 x 120)
